+++ backend/common/settings/structs.go
@@ -81,7 +81,7 @@ type SourceConfig struct {
 	IndexingInterval uint32             `json:"indexingIntervalMinutes"` // optional manual overide interval in seconds to re-index the source
 	DisableIndexing  bool               `json:"disableIndexing"`         // disable the indexing of this source
 	MaxWatchers      int                `json:"maxWatchers"`             // number of concurrent watchers to use for this source, currently not supported
-	NeverWatch       []string           `json:"neverWatchPaths"`         // paths to never watch, relative to the source path (eg. "/folder/file.txt")
+	NeverWatchPaths  []string           `json:"neverWatchPaths"`         // paths that get initially once. Useful for folders that rarely change contents (without source path prefix)
 	Exclude          ExcludeIndexFilter `json:"exclude"`                 // exclude files and folders from indexing, if include is not set
 	Include          IncludeIndexFilter `json:"include"`                 // include files and folders from indexing, if exclude is not set
 	DefaultUserScope string             `json:"defaultUserScope"`        // default "/" should match folders under path